revert ApplyUpdate changes, rename listener -> appdata
