+++ schema/appdata/listener.go
@@ -1,4 +1,4 @@
-package listener
+package appdata
 
 import (
 	"encoding/json"
